Add simple protocol RowConvertible
I originally tried to define `Row` as a protocol and have a number of
structs conforming to it, one per type or rows.

Unfortunately I can't seem to make that work... Seemed like I couldn't
use a concrete implementation of `Row` to init a `let x: [Row]` value.

For the sake of this first cut, having the wiring sorted is more
important than having a neat API. That will come later.