Improve how keyboard handling restores scroll view insets
This makes sure that the insets are restore to the original value rather
than an aribtratry zero.

This way the handling will work properly even if the view controller
containing the form is inside a navigation controller, or the form
itself has been modified by the user for some reason.