Only update bottom inset when displaying keyboard
The same way b40264d fixed how the scroll view insets are reset when the
keyboard is dismissed, this change makes sure that when the keyboard is
dispalyed we only adjust the inset to house the keyboard, not bluntly
set them to 0.

In practice this will result a text field at the start of the form not
going under the navigation bar anymore, because the top inset won't
become 0 but stay the current value set by the OS which already takes
into account the nav bar.