Add functions to notify FormController of VC lifecycle
These are needed to ensure that the keyboard handlers that resize the
form's table view are released properly when the view controller goes
out of screen. Otherwise every time a keyboard is shown or hidden and
the view controller that contains a `FormController` is still in memory,
it will adjust its table view's offset.